+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## v0.5 - 2020-
 
+## 0.4.18 - 2021-01-??
+- Added `packageID` variable to label, so version of packages can be found
+- `CheckInstallomator.sh` improved for version checking on github and can give warnings
+- Labels changed for Apple Silicon: slack
+- Labels with appNewVersion: slack
+- Labels changed: keyboardmaestro (bit better version checking), adobereaderdc & adobereaderdc-install & adobereaderdc-update (new appnewVersion), icons (error in downloadURL), vagrant (appNewVersion and fixed downloadURL)
+- Labels added: loom
+
 ## 0.4.17 - 2021-01-19
 - If any blocking process was closed, Installomator will try to open the app again, after the update process is done
 - Added a bit extra logging, when using INSTALL=force